+++ src/core/index.ts
@@ -1,6 +1,17 @@
 export { compileFile, compileSource } from "./compiler.js";
 export { parseIntentSource, ParserError } from "./parser.js";
-export { generateSwift, escapeSwiftString } from "./generator.js";
+export {
+  generateSwift,
+  generateInfoPlistFragment,
+  generateEntitlementsFragment,
+  escapeSwiftString,
+  escapeXml,
+} from "./generator.js";
 export { validateIntent, validateSwiftSource } from "./validator.js";
-export { irTypeToSwift, SWIFT_TYPE_MAP, PARAM_TYPES } from "./types.js";
+export {
+  irTypeToSwift,
+  SWIFT_TYPE_MAP,
+  PARAM_TYPES,
+  LEGACY_PARAM_ALIASES,
+} from "./types.js";
 export type * from "./types.js";